requirements.txt
pytest==7.4.0
requests==2.31.0
pyyaml==6.0.1